UNPC_Weighted_Sum
The goal in this problem is to sum all the digits in a number. However, each number must be “weighted” by its relative position in the number. For example, given the number 57 the weighted sum is 19. The weighted sum is calculated as: (5 * 1) + (7 * 2). Given the number 482178, the weighted sum is 113. The weighted sum is calculated as: (4 * 1) + (8 * 2) + (2 * 3) + (1 * 4) + (7 * 5) + (8 * 6).

Input Format

Read the positive integer number from the console.

Constraints

Only accept Numeric inputs.

Output Format

Display the weighted sum of the number. Display the output on the monitor. See the sample output below.

Sample Input 0

57
Sample Output 0

19